+++ macos_test.go
@@ -101,45 +101,6 @@ func TestUpdateItem(t *testing.T) {
 	}
 }
 
-func TestGenericPasswordRef(t *testing.T) {
-	service, account, label, accessGroup, password := "TestGenericPasswordRef", "test", "", "", "toomanysecrets"
-
-	item := NewGenericPassword(service, account, label, []byte(password), accessGroup)
-	defer func() { _ = DeleteItem(item) }()
-	err := AddItem(item)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Query reference and delete by reference
-	query := NewItem()
-	query.SetSecClass(SecClassGenericPassword)
-	query.SetService(service)
-	query.SetAccount(account)
-	query.SetMatchLimit(MatchLimitOne)
-	query.SetReturnRef(true)
-	ref, err := QueryItemRef(query)
-	if err != nil {
-		t.Fatal(err)
-	} else if ref == nil {
-		t.Fatal("Missing result")
-	} else {
-		err = DeleteItemRef(ref)
-		if err != nil {
-			t.Fatal(err)
-		}
-		Release(ref)
-	}
-
-	passwordAfter, err := GetGenericPassword(service, account, label, accessGroup)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if passwordAfter != nil {
-		t.Fatal("Shouldn't have password")
-	}
-}
-
 func TestAddingAndQueryingNewKeychain(t *testing.T) {
 	keychainPath := tempPath(t)
 	defer func() { _ = os.Remove(keychainPath) }()
